♻️ refactor(change_log_config): modify custom section type
- change Custom enum variant type from usize to u8 for optimized memory usage

+++ src/change_log_config.rs
@@ -51,7 +51,7 @@ pub enum DisplaySections {
     /// Display only the most recent section
     One,
     /// Display the lesser of the specified number and all sections
-    Custom(usize),
+    Custom(u8),
 }
 
 /// Pattern to identify a tag as a release tag.
